[UDT-192] feat: 설문조사 시작 화면 제작
